Carry every measure at the format's own width
Slap.Measure's byte quantities — Offset, Length, FileSize, Delta, SignedOffset, ParsedSizeValue — now wrap Int64, so a format's wire ranges stop depending on the host's word size; Position stays Int because it cursors an in-memory ByteString, and ActionIndex and SubstitutionCount stay Int as counts of in-memory things. offsetToInt, lengthToInt, and fileSizeToInt are the named narrowings at the ByteString boundary, sound only for values already bounded against a real buffer. ByteParser's fit checks now measure remaining room by subtraction rather than adding to the request, so a wire-wide count cannot wrap the comparison into a false pass, and NINJA1's record-end check keeps the buffer-realizability ceiling its apply fold trusts. The narrowing and rendering helpers (narrowToWord32 and kin, plural, commaNum, showSigned) take any integral rather than forcing conversions at their call sites.

+++ src/Slap/APSGBA/Create.hs
@@ -38,8 +38,8 @@ createAPSGBA inputContents@(InputFileContents original) outputContents@(OutputFi
     changedBlocks = filter hasChanges [0 .. blockCount - 1]
     hasChanges blockIndex =
       let offset = blockIndex * blockSize
-          sourceBlock = zeroExtendedBlock (Offset offset) (Length blockSize) original
-          targetBlock = zeroExtendedBlock (Offset offset) (Length blockSize) modified
+          sourceBlock = zeroExtendedBlock (Offset (fromIntegral offset)) (Length (fromIntegral blockSize)) original
+          targetBlock = zeroExtendedBlock (Offset (fromIntegral offset)) (Length (fromIntegral blockSize)) modified
       in sourceBlock /= targetBlock
 
 encodeGBABlock :: InputFileContents -> OutputFileContents -> Int -> Builder
@@ -53,6 +53,6 @@ encodeGBABlock (InputFileContents original) (OutputFileContents modified) blockI
     <> byteString xorPayload
   where
     offset = blockIndex * apsGbaBlockSize
-    sourceBlock = zeroExtendedBlock (Offset offset) (Length apsGbaBlockSize) original
-    targetBlock = zeroExtendedBlock (Offset offset) (Length apsGbaBlockSize) modified
+    sourceBlock = zeroExtendedBlock (Offset (fromIntegral offset)) (Length (fromIntegral apsGbaBlockSize)) original
+    targetBlock = zeroExtendedBlock (Offset (fromIntegral offset)) (Length (fromIntegral apsGbaBlockSize)) modified
     xorPayload = ByteString.packZipWith xor sourceBlock targetBlock
