+++ src/Slap/NINJA1/Apply.hs
@@ -21,7 +21,7 @@ import Data.Word (Word8)
 -- parse already refused any record reaching past 'Int' ('Slap.NINJA1.Parse.rejectUnaddressableRecordEnds'), so the arithmetic here cannot wrap.
 applyNINJA1 :: NINJA1Patch -> InputFileContents -> Either SlapError OutputFileContents
 applyNINJA1 patch (InputFileContents source) = Right $ OutputFileContents $ unsafeCreate outputSize $ \outputPointer -> do
-    copyRegion outputPointer (Offset 0) source (Offset 0) (Length (min sourceLength outputSize))
+    copyRegion outputPointer (Offset 0) source (Offset 0) (Length (fromIntegral (min sourceLength outputSize)))
     when (outputSize > sourceLength) $
       fillBytes (outputPointer `plusPtr` sourceLength) (0 :: Word8) (outputSize - sourceLength)
     forM_ (ninja1Records patch) $ \(NINJA1Record writeOffset writePayload) ->
